fixed search component
